[Refactor] 프로퍼티 관련 코드 분리
